requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pyusb==1.2.1
screeninfo==0.8.1
